#68 wordsUntil accept arrays of strings (cherry picked from commit e1627b8)
