+++ src/test/standard/extractor/extractor-bundle-test.js
@@ -294,6 +294,53 @@ export default {
 
         test.equals(value, 'World');
         test.done();
+    },
+
+    'test wordsUntilFast string': function (test) {
+
+        const line = stream.ofString('soXYZso');
+
+        const x = new X();
+        const combinator = x.wordsUntil('XYZ');
+        const parser = combinator.parse(line);
+        const value = parser.value;
+        const offset = parser.offset;
+
+        test.equals(value, 'so');
+        test.equals(offset, 2);
+        test.done();
+    },
+    'test wordsUntilFast string with continuation': function (test) {
+
+        const document = 'start-detect-XYZ-continues';
+        const line = stream.ofString(document);
+
+        const start=C.string('start-');
+        const x = new X();
+        const combinator = start.thenRight(x.wordsUntil('XYZ').thenLeft(C.string('XYZ-continues')));
+        const parser = combinator.parse(line);
+        const value = parser.value;
+        const offset = parser.offset;
+
+        test.equals(value, 'detect-');
+        test.equals(offset, document.length);
+        test.done();
+    },
+    'test wordsUntilFast array of string with continuation': function (test) {
+
+        const document = 'start-detect-XYZ-continues';
+        const line = stream.ofString(document);
+
+        const start=C.string('start-');
+        const x = new X();
+        const combinator = start.thenRight(x.wordsUntil(['ABC','ZE','XYZ']).thenLeft(C.string('XYZ-continues')));
+        const parser = combinator.parse(line);
+        const value = parser.value;
+        const offset = parser.offset;
+
+        test.equals(value, 'detect-');
+        test.equals(offset, document.length);
+        test.done();
     }
 
 }
